[FIX] warnings mainly in Gitlib

+++ src/Repository/Git/GitLuaGraphUtils.cpp
@@ -22,11 +22,11 @@ void Repository::Git::GitLuaGraphUtils::fillMetricsMetaData()
 		Lua::LuaNode* luaNode = iterator.value();
 
 		if ( luaNode->getParams()["type"].asString() == "function" ) {
-			int lines = luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines"].asNumber();
-			int linesBlank = luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_blank"].asNumber();
-			int linesCode = luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_code"].asNumber();
-			int linesComment = luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_comment"].asNumber();
-			int linesNonEmpty = luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_nonempty"].asNumber();
+			int lines = static_cast<int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines"].asNumber() );
+			int linesBlank = static_cast<int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_blank"].asNumber() );
+			int linesCode = static_cast<int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_code"].asNumber() );
+			int linesComment = static_cast<int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_comment"].asNumber() );
+			int linesNonEmpty = static_cast<int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_nonempty"].asNumber() );
 
 			this->evolutionGraph->getMetaDataFromIdentifier( luaNode->getIdentifier() )->insertMetrics( "lines", lines );
 			this->evolutionGraph->getMetaDataFromIdentifier( luaNode->getIdentifier() )->insertMetrics( "linesBlank", linesBlank );
@@ -43,11 +43,11 @@ QMap<QString, int>* Repository::Git::GitLuaGraphUtils::getMetrics( QString ident
 
 	Lua::LuaNode* luaNode = this->luaGraph->findNodeByLuaIdentifier( identifier );
 	if ( luaNode ) {
-		resultMap->insert( "lines", luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines"].asNumber() );
-		resultMap->insert( "linesBlank", luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_blank"].asNumber() );
-		resultMap->insert( "linesCode", luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_code"].asNumber() );
-		resultMap->insert( "linesComment", luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_comment"].asNumber() );
-		resultMap->insert( "linesNonEmpty", luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_nonempty"].asNumber() );
+		resultMap->insert( "lines", static_cast<const int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines"].asNumber() ) );
+		resultMap->insert( "linesBlank", static_cast<const int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_blank"].asNumber() ) );
+		resultMap->insert( "linesCode", static_cast<const int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_code"].asNumber() ) );
+		resultMap->insert( "linesComment", static_cast<const int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_comment"].asNumber() ) );
+		resultMap->insert( "linesNonEmpty", static_cast<const int>( luaNode->getParams()["metrics"].asTable()["LOC"].asTable()["lines_nonempty"].asNumber() ) );
 	}
 
 	return resultMap;
